Introduces SapDriver to manage computations specific to the SAP solver.
