+++ multibody/plant/contact_pair_kinematics.cc
@@ -1,0 +1,6 @@
+#include "drake/multibody/plant/contact_pair_kinematics.h"
+
+#include "drake/common/default_scalars.h"
+
+DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_SCALARS(
+    struct ::drake::multibody::internal::ContactPairKinematics)
