+++ multibody/plant/discrete_update_manager.cc
@@ -80,28 +80,11 @@ DiscreteUpdateManager<T>::EvalContactJacobians(
 }
 
 template <typename T>
-const std::vector<internal::DiscreteContactPair<T>>&
-DiscreteUpdateManager<T>::EvalDiscreteContactPairs(
-    const systems::Context<T>& context) const {
-  return MultibodyPlantDiscreteUpdateManagerAttorney<
-      T>::EvalDiscreteContactPairs(plant(), context);
-}
-
-template <typename T>
 const std::vector<geometry::ContactSurface<T>>&
 DiscreteUpdateManager<T>::EvalContactSurfaces(
     const systems::Context<T>& context) const {
   return MultibodyPlantDiscreteUpdateManagerAttorney<T>::EvalContactSurfaces(
       plant(), context);
-}
-
-template <typename T>
-std::vector<CoulombFriction<double>>
-DiscreteUpdateManager<T>::CalcCombinedFrictionCoefficients(
-    const systems::Context<T>& context,
-    const std::vector<internal::DiscreteContactPair<T>>& contact_pairs) const {
-  return MultibodyPlantDiscreteUpdateManagerAttorney<
-      T>::CalcCombinedFrictionCoefficients(plant(), context, contact_pairs);
 }
 
 template <typename T>
