+++ multibody/plant/multibody_plant_discrete_update_manager_attorney.h
@@ -53,21 +53,9 @@ class MultibodyPlantDiscreteUpdateManagerAttorney {
     return plant.EvalContactJacobians(context);
   }
 
-  static const std::vector<internal::DiscreteContactPair<T>>&
-  EvalDiscreteContactPairs(const MultibodyPlant<T>& plant,
-                           const systems::Context<T>& context) {
-    return plant.EvalDiscreteContactPairs(context);
-  }
-
   static const std::vector<geometry::ContactSurface<T>>& EvalContactSurfaces(
       const MultibodyPlant<T>& plant, const systems::Context<T>& context) {
     return plant.EvalContactSurfaces(context);
-  }
-
-  static std::vector<CoulombFriction<double>> CalcCombinedFrictionCoefficients(
-      const MultibodyPlant<T>& plant, const systems::Context<T>& context,
-      const std::vector<internal::DiscreteContactPair<T>>& contact_pairs) {
-    return plant.CalcCombinedFrictionCoefficients(context, contact_pairs);
   }
 
   static void AddInForcesFromInputPorts(const MultibodyPlant<T>& plant,
